dhcp: use "nettools" DHCPv4 implementation instead of systemd's for "internal" plugin
Note that there are still undocumented settings for "main.dhcp" in "NetworkManager.conf"
to explicitly select "systemd" or "nettools". That is a fallback during the testing
phase, if the "nettools" plugin has a serious issue. These are not
supposed to be officially supported plugins. Officially there is only
the "internal" plugin, which as of now uses nettools's n-dhcp4 for IPv4.